test(clickclack): isolate timeout regressions
